refactor : 스페이스 생성 로직을 SpaceService 클래스로 분리
- 유저의 개인 및 단체 스페이스를 초기화하고 저장하는 작업을 비동기적으로 처리
