+++ src/main/java/com/javajober/member/service/MemberService.java
@@ -1,9 +1,7 @@
 package com.javajober.member.service;
 
 
-import java.util.HashSet;
 import java.util.Optional;
-import java.util.Set;
 
 import javax.transaction.Transactional;
 
@@ -21,10 +19,9 @@ import com.javajober.member.repository.MemberRepository;
 import com.javajober.core.refreshToken.repository.RefreshTokenRepository;
 import com.javajober.core.security.JwtTokenizer;
 import com.javajober.core.refreshToken.domain.RefreshToken;
-import com.javajober.space.domain.AddSpace;
-import com.javajober.space.domain.SpaceType;
-import com.javajober.space.dto.request.SpaceSaveRequest;
-import com.javajober.space.repository.AddSpaceRepository;
+import com.javajober.space.service.SpaceService;
+
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 public class MemberService {
@@ -32,15 +29,15 @@ public class MemberService {
 	private final PasswordEncoder passwordEncoder;
 	private final JwtTokenizer jwtTokenizer;
 	private final RefreshTokenRepository refreshTokenRepository;
-	private final AddSpaceRepository addSpaceRepository;
+	private final SpaceService spaceService;
 
 	public MemberService(MemberRepository memberRepository, PasswordEncoder passwordEncoder, JwtTokenizer jwtTokenizer,
-		RefreshTokenRepository refreshTokenRepository, AddSpaceRepository addSpaceRepository) {
+		RefreshTokenRepository refreshTokenRepository, SpaceService spaceService) {
 		this.memberRepository = memberRepository;
 		this.passwordEncoder = passwordEncoder;
 		this.jwtTokenizer = jwtTokenizer;
 		this.refreshTokenRepository = refreshTokenRepository;
-		this.addSpaceRepository = addSpaceRepository;
+		this.spaceService = spaceService;
 	}
 
 	@Transactional
@@ -55,37 +52,9 @@ public class MemberService {
 		member.setPassword(passwordEncoder.encode(memberSignupRequest.getPassword()));
 		Member saveMember = memberRepository.save(member);
 
-		initializeAndSaveNewMemberSpaces(member);
+		spaceService.initializeAndSaveNewMemberSpaces(member);
 
 		return new MemberSignupResponse(saveMember);
-	}
-
-	private void initializeAndSaveNewMemberSpaces(Member member) {
-
-		SpaceSaveRequest personalSpaceRequest = createSpaceSaveRequest(member.getMemberName(), SpaceType.PERSONAL.getEngTitle(), member.getMemberName());
-		SpaceSaveRequest organizationSpaceRequest = createSpaceSaveRequest(member.getMemberName(), SpaceType.ORGANIZATION.getEngTitle(), "임시회사명");
-
-		Set<AddSpace> spaces = new HashSet<>();
-
-		AddSpace personalSpace = SpaceSaveRequest.toEntity(personalSpaceRequest, member);
-		spaces.add(personalSpace);
-
-		AddSpace organizationSpace = SpaceSaveRequest.toEntity(organizationSpaceRequest, member);
-		spaces.add(organizationSpace);
-
-		saveSpaces(spaces);
-	}
-
-	private SpaceSaveRequest createSpaceSaveRequest(String spaceTitle, String spaceType, String representativeName) {
-			return SpaceSaveRequest.builder()
-				.spaceTitle(spaceTitle)
-				.spaceType(spaceType)
-				.representativeName(representativeName)
-				.build();
-	}
-
-	private void saveSpaces(Set<AddSpace> spaces) {
-		addSpaceRepository.saveAll(spaces);
 	}
 
 	@Transactional
